backend/requirements.txt
<<<<<<< HEAD
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.32.3
groq==0.9.0
=======
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
groq
>>>>>>> eade00dfe17ef9cfe30c002f6a61978f4f720ec6
